Moved NetworkMonitor to NetworkPlugin
